# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for asset management system.",
        "title": "Asset Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:6500",
    "paths": {
...
        "/sys/excel-field-match/create": {
            "post": {
                "description": "Create ExcelFieldMatch record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Excel Field Match"
                ],
                "summary": "Create ExcelFieldMatch record",
                "parameters": [
                    {
                        "description": "Excel Field Match Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExcelFieldMatchs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralCreateResponseBody"
                        }
                    }
                }
            }
        },
        "/sys/excel-field-match/list": {
            "post": {
                "description": "List ExcelFieldMatch records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Excel Field Match"
                ],
                "summary": "List ExcelFieldMatch records",
                "parameters": [
                    {
                        "description": "List ExcelFieldMatch Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ExcelFieldMatchPageDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ExcelFieldMatchList"
                        }
                    }
                }
            }
        },
        "/sys/excel-field-match/one/{id}": {
            "get": {
                "description": "Get one ExcelFieldMatch record by code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Excel Field Match"
                ],
                "summary": "Get one ExcelFieldMatch record by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExcelFieldMatchs"
                        }
                    }
                }
            }
        },
        "/sys/excel-field-match/update": {
            "post": {
                "description": "Update ExcelFieldMatch record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Excel Field Match"
                ],
                "summary": "Update ExcelFieldMatch record",
                "parameters": [
                    {
                        "description": "Excel Field Match Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExcelFieldMatchs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateResponse"
                        }
                    }
                }
            }
        },
        "/sys/excel-field-match/void/{id}": {
            "delete": {
                "description": "Void one ExcelFieldMatch record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Excel Field Match"
                ],
                "summary": "Void one ExcelFieldMatch record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ExcelFieldMatch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateResponse"
                        }
                    }
                }
            }
        },
        "/sys/inv-record/list": {
            "post": {
                "description": "List Inventory Records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory Record"
                ],
                "summary": "List Inventory Records",
                "parameters": [
                    {
                        "description": "List Action Record Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InvRecordListDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Inventory Records",
                        "schema": {
                            "$ref": "#/definitions/dto.InvRecordListResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ExcelFieldMatchList": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExcelFieldMatchs"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.ExcelFieldMatchPageDto": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralCreateResponseBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dto.GeneralUpdateResponse": {
            "type": "object",
            "properties": {
                "MatchedCount": {
                    "type": "integer"
                },
                "ModifiedCount": {
                    "type": "integer"
                },
                "UpsertedCount": {
                    "type": "integer"
                },
                "UpsertedID": {
                    "type": "string"
                }
            }
        },
...
        "dto.InvRecordList": {
            "type": "object",
            "properties": {
                "assetCode": {
                    "type": "string"
                },
                "assetList": {
                    "$ref": "#/definitions/models.AssetLists"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "placeFrom": {
                    "type": "string"
                },
                "placeFromData": {
                    "$ref": "#/definitions/models.Locations"
                },
                "placeTo": {
                    "type": "string"
                },
                "placeToData": {
                    "$ref": "#/definitions/models.Locations"
                }
            }
        },
        "dto.InvRecordListDto": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "dto.InvRecordListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InvRecordList"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.AssetListFiles": {
            "type": "object",
            "properties": {
                "assetId": {
                    "type": "string"
                },
                "base64": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AssetLists": {
            "type": "object",
            "properties": {
                "accountCode": {
                    "type": "string"
                },
                "accountName": {
                    "type": "string"
                },
                "afterBeforeTax": {
                    "type": "number"
                },
                "assetCode": {
                    "type": "string"
                },
                "assetName": {
                    "type": "string"
                },
                "brandCode": {
                    "type": "string"
                },
                "brandName": {
                    "type": "string"
                },
                "chequeNo": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deptId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "includeTax": {
                    "type": "boolean"
                },
                "invoiceDate": {
                    "type": "string"
                },
                "invoiceNo": {
                    "type": "string"
                },
                "invoiceRemark": {
                    "description": "I",
                    "type": "string"
                },
                "maintenancePeriodEnd": {
                    "type": "string"
                },
                "maintenancePeriodStart": {
                    "type": "string"
                },
                "placeId": {
                    "type": "string"
                },
                "purchaseDate": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "serialNo": {
                    "type": "string"
                },
                "sponsor": {
                    "type": "boolean"
                },
                "sponsorName": {
                    "type": "string"
                },
                "staffName": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "taxCode": {
                    "type": "string"
                },
                "taxCountryCode": {
                    "type": "string"
                },
                "taxInfoId": {
                    "type": "string"
                },
                "taxRate": {
                    "type": "number"
                },
                "typeId": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uploadAssetListFiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AssetListFiles"
                    }
                },
                "vendorId": {
                    "type": "string"
                },
                "voucherNo": {
                    "type": "string"
                },
                "voucherUsedDate": {
                    "type": "string"
                }
            }
        },
...
        "models.ExcelFieldMatchs": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fieldLists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FieldListsModel"
                    }
                },
                "functionCode": {
                    "type": "string"
                },
                "functionName": {
                    "type": "string"
                },
                "functionType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.FieldListsModel": {
            "type": "object",
            "properties": {
                "dbFieldName": {
                    "type": "string"
                },
                "excelFieldName": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "models.Locations": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "placeCode": {
                    "type": "string"
                },
                "placeName": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}